+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.5 -- themes, css, injectGlobal
+
+1. themes via (setTheme), makes globally available in state
+2. css function now available
+3. injectGlobal for global css injection
+
 # 1.0.1 -- keyframes and optional double method notation
 
 1. Supports @keyframes now via styled.keyframes